RF Add base page for common methods in the MainPage, MyViewPage and ViewPage
