+++ src/test/java/school/redrover/PipelineTest.java
@@ -29,7 +29,7 @@ public class PipelineTest extends BaseTest {
                 .clickSaveButton()
                 .getHeader()
                 .clickLogo()
-                .getJobName();
+                .getJobName(NAME);
 
         Assert.assertEquals(projectName, NAME);
     }
@@ -116,7 +116,7 @@ public class PipelineTest extends BaseTest {
                 .clickRenameButton()
                 .getHeader()
                 .clickLogo()
-                .getJobName();
+                .getJobName(NEW_NAME);
 
         Assert.assertEquals(projectName, NEW_NAME);
     }
@@ -254,7 +254,7 @@ public class PipelineTest extends BaseTest {
                 .selectScriptedPipelineAndSubmit()
                 .getHeader()
                 .clickLogo()
-                .verifyJobIsPresent(NAME);
+                .jobIsDisplayed(NAME);
 
         Assert.assertTrue(projectIsPresent);
     }
@@ -319,7 +319,7 @@ public class PipelineTest extends BaseTest {
                 .clickRenameButton()
                 .getHeader()
                 .clickLogo()
-                .getJobName();
+                .getJobName(NEW_NAME);
 
         Assert.assertEquals(renamedPipeline, NEW_NAME);
     }
@@ -336,7 +336,7 @@ public class PipelineTest extends BaseTest {
                 .clickSaveButton()
                 .getHeader()
                 .clickLogo()
-                .getProjectNameMainPage(allowedChar);
+                .getJobName(allowedChar);
 
         Assert.assertEquals(projectNameDashboard, allowedChar);
     }
@@ -488,7 +488,7 @@ public class PipelineTest extends BaseTest {
 
         Assert.assertEquals(pipelinePage.getJobName(), "Pipeline " + NEW_NAME);
         Assert.assertEquals(pipelinePage.getProjectNameSubtitleWithDisplayName(), NAME);
-        Assert.assertEquals(pipelinePage.getHeader().clickLogo().getJobName(), NEW_NAME);
+        Assert.assertEquals(pipelinePage.getHeader().clickLogo().getJobName(NEW_NAME), NEW_NAME);
     }
 
     @Test
@@ -563,7 +563,7 @@ public class PipelineTest extends BaseTest {
         boolean projectIsPresent = new MainPage(getDriver())
                 .dropDownMenuClickDelete(NEW_NAME)
                 .dismissAlert()
-                .verifyJobIsPresent(NEW_NAME);
+                .jobIsDisplayed(NEW_NAME);
 
         Assert.assertTrue(projectIsPresent);
     }
