+++ src/test/java/school/redrover/UsersTest.java
@@ -277,7 +277,7 @@ public class UsersTest extends BaseTest {
                 .enterPassword(PASSWORD)
                 .enterSignIn(new MainPage(getDriver()));
         TestUtils.createJob(this, nameProject, TestUtils.JobType.FreestyleProject, true);
-        String actualResult = new MainPage(getDriver()).getJobName();
+        String actualResult = new MainPage(getDriver()).getJobName(nameProject);
 
         Assert.assertEquals(actualResult, nameProject);
     }
