+++ src/test/java/school/redrover/model/MyViewsPage.java
@@ -4,36 +4,13 @@ import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.FindBy;
 import org.openqa.selenium.support.ui.ExpectedConditions;
-import school.redrover.model.base.BaseMainHeaderPage;
+import school.redrover.model.base.BaseDashboardPage;
 import school.redrover.runner.TestUtils;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class MyViewsPage extends BaseMainHeaderPage<MyViewsPage> {
-    @FindBy(css = "#ok-button")
-    private WebElement okButton;
-
-    @FindBy(id = "description-link")
-    private WebElement onDescription;
-
-    @FindBy(xpath = "//textarea[@name='description']")
-    private WebElement enterDescription;
-
-    @FindBy(xpath = "//button[@class='jenkins-button jenkins-button--primary ']")
-    private WebElement saveButtonDescription;
-
-    @FindBy(xpath = "//textarea[@name='description']")
-    private WebElement textFromDescription;
-
-    @FindBy(xpath = "//*[@id=\"description\"]/div[1]")
-    private WebElement descriptionText;
-
-    @FindBy(css = ".task-link-wrapper>a[href$='newJob']")
-    private WebElement newJobNewItem;
-
-    @FindBy(xpath = "//a[@class='addTab']")
-    private WebElement newViewButton;
+public class MyViewsPage extends BaseDashboardPage<MyViewsPage> {
 
     public MyViewsPage(WebDriver driver) {
         super(driver);
@@ -51,9 +28,6 @@ public class MyViewsPage extends BaseMainHeaderPage<MyViewsPage> {
     @FindBy(xpath = "//h2")
     private WebElement statusMessage;
 
-    @FindBy(xpath = "//div[@class = 'tab active']")
-    private WebElement activeView;
-
     @FindBy(xpath = "//div[@class = 'tab'][last()-1]")
     private WebElement inactiveLastCreatedMyView;
 
@@ -63,72 +37,9 @@ public class MyViewsPage extends BaseMainHeaderPage<MyViewsPage> {
     @FindBy(xpath = "//div[@class='tabBar']//div[starts-with(@class, 'tab')]")
     private List<WebElement> allViews;
 
-    @FindBy(xpath = "//a[@href='newJob']")
-    private WebElement createAJob;
-
-    @FindBy(id = "name")
-    private WebElement anItemName;
-
-    @FindBy(css = ".hudson_model_FreeStyleProject")
-    private WebElement freestyleProject;
-
-    @FindBy(xpath = "//button[@formnovalidate = 'formNoValidate']")
-    private WebElement saveButton;
-
-    public MyViewsPage clickOkButton() {
-        getWait5().until(ExpectedConditions.visibilityOf(okButton)).click();
-        return this;
-    }
-
-    public MyViewsPage clickOnDescription() {
-        getWait5().until(ExpectedConditions.visibilityOf(onDescription)).click();
-        return new MyViewsPage(getDriver());
-    }
-
-    public MyViewsPage enterDescription(String name) {
-        getWait5().until(ExpectedConditions.visibilityOf(enterDescription)).sendKeys(name);
-        return new MyViewsPage(getDriver());
-    }
-
-    public MyViewsPage clickSaveButtonDescription() {
-        getWait5().until(ExpectedConditions.visibilityOf(saveButtonDescription)).click();
-        return new MyViewsPage(getDriver());
-    }
-
-    public String getTextFromDescription() {
-
-        return getWait5().until(ExpectedConditions.visibilityOf(descriptionText)).getText();
-    }
-
-    public MyViewsPage clearTextFromDescription() {
-        getWait10().until(ExpectedConditions.visibilityOf(textFromDescription)).clear();
-        return new MyViewsPage(getDriver());
-    }
-
-    public MyViewsPage enterNewDescription(String name) {
-        getWait5().until(ExpectedConditions.visibilityOf(textFromDescription)).sendKeys(name);
-        return new MyViewsPage(getDriver());
-    }
-
-    public NewJobPage clickNewItem() {
-        newJobNewItem.click();
-        return new NewJobPage(getDriver());
-    }
-
     public String getStatusMessageText() {
 
         return statusMessage.getText();
-    }
-
-    public NewViewPage clickNewViewButton() {
-        TestUtils.click(this, newViewButton);
-
-        return new NewViewPage(getDriver());
-    }
-
-    public String getActiveView() {
-
-        return TestUtils.getText(this, activeView);
     }
 
     public MyViewsPage clickInactiveLastCreatedMyView() {
@@ -159,26 +70,5 @@ public class MyViewsPage extends BaseMainHeaderPage<MyViewsPage> {
         }
 
         return list;
-    }
-
-    public MyViewsPage clickCreateAJob() {
-        getWait5().until(ExpectedConditions.visibilityOf(createAJob)).click();
-        return new MyViewsPage(getDriver());
-    }
-
-    public MyViewsPage enterAnItemName(String name) {
-        getWait5().until(ExpectedConditions.visibilityOf(anItemName)).sendKeys(name);
-        return this;
-    }
-
-    public MyViewsPage clickFreestyleProject() {
-        getWait5().until(ExpectedConditions.visibilityOf(freestyleProject)).click();
-        return this;
-    }
-
-    public MyViewsPage clickSaveButton() {
-        getWait5().until(ExpectedConditions.visibilityOf(saveButton)).click();
-
-        return this;
     }
 }
